hf-space/requirements.txt
torch==2.6.0
torchvision==0.21.0
timm>=1.0.0
Pillow>=10.0
huggingface_hub>=0.20.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
gradio>=5.0.0